Rework  Sample1 so that StateData is no longer used (or needed)
This was done by turning the namespace PlayerHsm into an inner class of Player (the owner) so that the states, which are themselves inner classes of PlayerHsm now, can access private data on the owner class. This effetively takes advantage of the fact that inner types have access to an outer type's privates (recursively).